Fix using no-cache option for the container build
If we specify no-cache, we should not add the local cache with the
from-cache and to-cache parameters. Otherwise, we get the error

```
WARNING: local cache import at /home/adminuser/.holoscan_build_cache
not found due to err: could not read
/home/adminuser/.holoscan_build_cache/index.json: open
/home/adminuser/.holoscan_build_cache/index.json: no such file or directory
```
being adminuser the user that runs the build.

This is important for CI, where we do not have any cache to start with.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.com>